"Initialize Scrapy project for image scraping
We have added the configuration files and code necessary to create a Scrapy project to scrape images. The spider "google_images_spider" which will scrape images from Google Images is also included. The project fetches the images using the downloaded URLs and stores them in the local 'downloads' directory. This was done to automate the process of bulk image download, which will be beneficial for dataset creation in machine learning tasks."