Installed Selenium webdriver to modify image scraper
The scraper now allows the use of Selenium within the image scraping process. A new middleware class, SeleniumMiddleware, was created to process requests using a Selenium webdriver. The decision to integrate Selenium comes from the need to effectively handle JavaScript on the page, making it possible to scrape dynamically loaded images. To avoid request routing through proxies when making requests for images, a 'dont_proxy' field is added to the request meta-data in the get_media_requests function in the MyImagesPipeline class. Also, Selenium was added in the requirements.txt file.