Added Dockerfile for creating a Docker image for the scraper
This commit includes the addition of a Dockerfile to enable the creation of Docker images for the image_scraper project. This will allow the scraper to be containerized and run in any environment that supports Docker. Also, a `*.http` pattern has been added to the .gitignore file to ignore HTTP Client request files that should not be included in the codebase.