Add Docker configuration for Scrapyd server and deployment
Added Docker configuration files to support deployment and running of the Scrapyd server in a container. Dockerizing the server would simplify the deployment process and ensure all the dependencies are met in a consistent manner. The Python runtime version was also updated to 3.10.

Furthermore, created a distinct Dockerfile for the scrapyd-deploy service and a new docker-compose file for ease of managing these services. Also added chromium and ChromeDriver installation to the Dockerfile to support web scraping in headless browser mode. Updated scrapyd URL in the scrapy.cfg file pointing it to the new scrapyd service.

This change mainly aims to make the image scraping service more portable, scalable, and easy to deploy.