Refactor image scraper for increased efficiency and clarity
A driver instance has been made class-specific which eliminates the need for constant reassignment at each usage. number of scrolls has been increased to fetch more images. Additionally, repetitive piece of code, that fetches and validates image url is replaced with a function scrape_image_url, making the code DRY-compliant. In the pipeline, a unique image name is generated using a hash of the URL to avoid overwriting images with the same name. This refactor aims to improve the scraping speed, code readability, and management of the downloaded images.