Remove Airflow network dependency from Docker Compose
This commit removes the 'airflow_default' network references from the 'docker-compose.yaml' file within the 'image_scraper' tool. Additionally, the project name 'mlist' is specified in the Makefile commands to streamline the Docker Compose usage. These changes were necessary as the 'airflow_default' network was not required for the operation of the respective services and introducing the project name enables the effective segregation of resources.